% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agTrend-package.R
\docType{data}
\name{edpsNonpups}
\alias{edpsNonpups}
\title{Pup counts from surveys of the eastern Distict Populations Segment (eDPS) of Steller sea lions.}
\format{A data frame with 86 observations on the following 3 variables.

\describe{ 
\item{region}{One of 4 eDPS subregions: southeast Alaska (SE AK), British Columbia (BC), Oregon (OR), and California (CA)}

\item{year}{Year the survey was conducted}

\item{count}{Aggregated count of pups of sites within each subregion.}
}}
\source{
Alaska Ecosystems Program, Marine Mammal Laboratory, Alaska
Fisheries Science Center, National Marine Fisheries Service, NOAA, 7600 Sand
Point Way, NE Seattle, WA 98115
}
\description{
Pup counts from surveys of the eastern Distict Populations Segment (eDPS) of Steller sea lions.
}
\examples{

data(edpsPups)
head(edpsPups)
}
\keyword{datasets}
