% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{edpsPups}
\alias{edpsPups}
\title{Pup counts from surveys of the eastern Distict Populations Segment (eDPS) of Steller sea lions.}
\format{A data frame with 35 observations on the following 3 variables.

\describe{
\item{region}{One of 4 eDPS subregions: southeast Alaska (SE AK), British Columbia (BC), Oregon (OR), and California (CA)}

\item{year}{Year the survey was conducted}

\item{count}{Aggregated count of pups of sites within each subregion.}
}}
\source{
Alaska Ecosystems Program, National Marine Mammal Laboratory, Alaska
Fisheries Science Center, National Marine Fisheries Service, NOAA, 7600 Sand
Point Way, NE Seattle, WA 98115
}
\description{
Pup counts from surveys of the eastern Distict Populations Segment (eDPS) of Steller sea lions.
}
\examples{
data(edpsPups)
head(edpsPups)
}
\keyword{datasets}

