\docType{data}
\name{photoCorrection}
\alias{photoCorrection}
\title{Data from a pilot study of collecting Steller sea lion counts from photos taken by hand at
oblique angles vs. vertical medium-format photos.}
\format{A data frame with 20 observations on the following 4 variables.

\describe{ \item{site}{Site where survey was taken}

\item{Rookery}{1 if site is a rookery, 0 if else.}

\item{Region}{Region in which the site is located}

\item{X2000OBL}{Survey count taken from oblique photo source}

\item{X2000VERT}{Survey count taken from vertical medium-format photo source}

}}
\source{
  Alaska Ecosystems Program, National Marine Mammal
  Laboratory, Alaska Fisheries Science Center, National
  Marine Fisheries Service, NOAA, 7600 Sand Point Way, NE
  Seattle, WA 98115
}
\description{
  Data from a pilot study of collecting Steller sea lion
  counts from photos taken by hand at oblique angles vs.
  vertical medium-format photos.
}
\examples{
data(photCorrection)
head(photoCorrection)
}
\keyword{datasets}

