% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agTrend-package.R
\docType{data}
\name{wdpsNonpups}
\alias{wdpsNonpups}
\title{Steller sea lion survey data (nonpup counts) collected by the NOAA National Marine Mammal Laboratory in the western Distinct 
Population Segment of the stock (wDPS).}
\format{A data frame with 4118 observations on the following 5 variables.

\describe{ \item{Site}{Site where survey was taken}

\item{Region}{Region in which the site is located}

\item{RCA}{Rookery Cluster Area in which the site belongs}

\item{year}{Year in which the survey was conducted}

\item{count}{The count of nonpups observed}
}}
\source{
Alaska Ecosystems Program, Marine Mammal Laboratory, Alaska
Fisheries Science Center, National Marine Fisheries Service, NOAA, 7600 Sand
Point Way, NE Seattle, WA 98115
}
\description{
Steller sea lion survey data (nonpup counts) collected by the NOAA National Marine Mammal Laboratory in the western Distinct 
Population Segment of the stock (wDPS).
}
\examples{

data(wdpsNonpups)
head(wdpsNonpups)
}
\references{
Need a tech report / memo reference...
}
\keyword{datasets}
