% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agTrend.R
\name{updateTrend}
\alias{updateTrend}
\title{Recalculate posterior predictive trend with a different time scale.}
\usage{
updateTrend(x, start, end, type = "pred", order = "lin")
}
\arguments{
\item{x}{An mcmc augmentation object produced by a call to \code{\link{mcmc.aggregate}} or 
an element of the list produced by a call to \code{\link{newAggregation}}.}

\item{start}{A new start value for the time span}

\item{end}{A new end value for the time span}

\item{type}{The type of trend calculated. Use \code{"pred"} for posterior predictive trends
and \code{"real"} to use the estimated, realized abumndance aggregation.}

\item{order}{The order of trend calculated. Can be one of \code{"lin"}, for linear trends, 
or, \code{"const"}, for mean log-abundence.}
}
\description{
Given a fitted model object (fit with 'keep.site.abund=TRUE') the function recalculates the posterior sample for an alternate 
time scale than was originally specified in the model fitting MCMC sample.
}
