% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agTrend-package.R
\docType{data}
\name{wdpsPups}
\alias{wdpsPups}
\title{Pup counts from surveys of the western Distict Populations Segment (wDPS) of Steller sea lions.}
\format{A data frame with 694 observations on the following 5 variables.

\describe{ 
\item{Site}{Surveyed sites in the wDPS}

\item{Region}{Region of the surveyed site}

\item{RCA}{Rookery cluster area of the surveyed site}

\item{year}{Survey year}

\item{count}{Survey count of pups at each site}
}}
\source{
Alaska Ecosystems Program, National Marine Mammal Laboratory, Alaska
Fisheries Science Center, National Marine Fisheries Service, NOAA, 7600 Sand
Point Way, NE Seattle, WA 98115
}
\description{
Pup counts from surveys of the western Distict Populations Segment (wDPS) of Steller sea lions.
}
\examples{

data(wdpsPups)
head(wdpsPups)
}
\keyword{datasets}
